+++ src/components/CampaignsAdmin/DropPhotoZone.tsx
